High ping fix
This offers an option for players with high ping to eliminate the accelmeter display latency - meaning that it appears to them as if they had a ping of zero.

Player.buttons (the cause of the latency) is bypassed. Some old code that was originally commented out is used for this - the code is a bit buggy (which is why player.buttons is used instead), but it nevertheless is preferable for someone with high ping.